requirements.txt
pymupdf==1.25.0
pdfplumber==0.10.3
easyocr==1.7.1
pytesseract==0.3.10
Pillow==10.2.0
pandas==2.1.4
numpy==1.26.3
camelot-py==0.11.0
opencv-python==4.9.0.80
tabula-py==2.9.0
